+++ gst/gst-kvs-plugin/src/KvsWebRtc.h
@@ -41,8 +41,6 @@
 #define IS_NALU_H265_IDR_HEADER(h)         ((((h) >> 1) == IDR_W_RADL_NALU_TYPE || ((h) >> 1) == IDR_N_LP_NALU_TYPE))
 #define IS_NALU_H265_VPS_SPS_PPS_HEADER(h) (((h) >> 1) == H265_VPS_NALU_TYPE || ((h) >> 1) == H265_SPS_NALU_TYPE || ((h) >> 1) == H265_PPS_NALU_TYPE)
 
-#include "View.h"
-
 typedef VOID (*StreamSessionShutdownCallback)(UINT64, PWebRtcStreamingSession);
 
 STATUS signalingClientStateChangedFn(UINT64, SIGNALING_CLIENT_STATE);
